Add Basic Ship Functionalities and Space Backdrop
Add the ability for the ship to fire projectiles when there's an obstacle in its path. These projectiles should be children of the level so if the ship dies, they aren't deleted.

Add a sphere to the world scene and give it a scrolling shader so there's a space effect in the background.